EquipmentInfo double to decimal refactoring

+++ Server/DealnetPortal.Domain/EquipmentInfo.cs
@@ -35,13 +35,13 @@ namespace DealnetPortal.Domain
 
         public DeferralType? DeferralType { get; set; }
 
-        public double? CustomerRate { get; set; }
+        public decimal? CustomerRate { get; set; }
         
-        public double? AdminFee { get; set; }
+        public decimal? AdminFee { get; set; }
         
-        public double? DownPayment { get; set; }
+        public decimal? DownPayment { get; set; }
 
-        public double? ValueOfDeal { get; set; }
+        public decimal? ValueOfDeal { get; set; }
 
         public string SalesRep { get; set; }
 
@@ -82,6 +82,6 @@ namespace DealnetPortal.Domain
 
         public Contract Contract { get; set; }
 
-        public double? DealerCost { get; set; }
+        public decimal? DealerCost { get; set; }
     }
 }
